+++ model/model.ts
@@ -1,9 +1,11 @@
 export interface Task {
   id: string;
   description: string;
-  completed: boolean;
-
+  
   duration: number;
   start: number;
   end: number;
+  
+  completed: boolean;
+  priority: number
 }